Make gateway read-only by default and add option to make it writable

+++ core/corehttp/gateway_handler.go
@@ -41,13 +41,15 @@ type directoryItem struct {
 // gatewayHandler is a HTTP handler that serves IPFS objects (accessible by default at /ipfs/<path>)
 // (it serves requests like GET /ipfs/QmVRzPKPzNtSrEzBFm2UZfxmPAgnaLke4DMcerbsGGSaFe/link)
 type gatewayHandler struct {
-	node    *core.IpfsNode
-	dirList *template.Template
+	node     *core.IpfsNode
+	dirList  *template.Template
+	writable bool
 }
 
-func newGatewayHandler(node *core.IpfsNode) (*gatewayHandler, error) {
+func newGatewayHandler(node *core.IpfsNode, writable bool) (*gatewayHandler, error) {
 	i := &gatewayHandler{
-		node: node,
+		node:     node,
+		writable: writable,
 	}
 	err := i.loadTemplate()
 	if err != nil {
@@ -90,18 +92,25 @@ func (i *gatewayHandler) NewDagReader(nd *dag.Node) (io.Reader, error) {
 func (i *gatewayHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	path := r.URL.Path[5:]
 
-	if r.Method == "POST" && path == "/" {
+	if i.writable && r.Method == "POST" && path == "/" {
 		i.postHandler(w, r)
 		return
 	}
 
-	if r.Method == "PUT" {
+	if i.writable && r.Method == "PUT" {
 		i.putHandler(w, r, path)
 		return
 	}
 
-	if r.Method == "DELETE" {
+	if i.writable && r.Method == "DELETE" {
 		i.deleteHandler(w, r, path)
+		return
+	}
+
+	if r.Method != "GET" {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("Method " + r.Method + " not allowed: bad request or read only access"))
+		log.Error("Method %s not allowed: bad request or read only access")
 		return
 	}
